+++ src/Symfony/Component/Validator/Tests/Constraints/ExpressionSyntaxValidatorTest.php
@@ -15,6 +15,7 @@ use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 use Symfony\Component\Validator\Constraints\ExpressionSyntax;
 use Symfony\Component\Validator\Constraints\ExpressionSyntaxValidator;
 use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
+use Symfony\Component\Validator\Tests\Constraints\Fixtures\StringableValue;
 
 class ExpressionSyntaxValidatorTest extends ConstraintValidatorTestCase
 {
@@ -47,6 +48,16 @@ class ExpressionSyntaxValidatorTest extends ConstraintValidatorTestCase
         $this->assertNoViolation();
     }
 
+    public function testStringableExpressionValid()
+    {
+        $this->validator->validate(new StringableValue('1 + 1'), new ExpressionSyntax([
+            'message' => 'myMessage',
+            'allowedVariables' => [],
+        ]));
+
+        $this->assertNoViolation();
+    }
+
     public function testExpressionWithoutNames()
     {
         $this->validator->validate('1 + 1', new ExpressionSyntax([
@@ -69,6 +80,20 @@ class ExpressionSyntaxValidatorTest extends ConstraintValidatorTestCase
     public function testExpressionIsNotValid()
     {
         $this->validator->validate('a + 1', new ExpressionSyntax([
+            'message' => 'myMessage',
+            'allowedVariables' => [],
+        ]));
+
+        $this->buildViolation('myMessage')
+            ->setParameter('{{ syntax_error }}', '"Variable "a" is not valid around position 1 for expression `a + 1`."')
+            ->setInvalidValue('a + 1')
+            ->setCode(ExpressionSyntax::EXPRESSION_SYNTAX_ERROR)
+            ->assertRaised();
+    }
+
+    public function testStringableExpressionIsNotValid()
+    {
+        $this->validator->validate(new StringableValue('a + 1'), new ExpressionSyntax([
             'message' => 'myMessage',
             'allowedVariables' => [],
         ]));